airflow/requirements.txt
pandas==1.3.3
requests==2.27.1
python-dotenv==0.20.0
SQLAlchemy==1.4.9
sqlalchemy-redshift==0.8.9
redshift-connector==2.0.908
psycopg2-binary==2.9.3